[PLINK-91] - Move the Credentials related classed from the IMPL to the API module
