+++ idm/api/pom.xml
@@ -1,53 +1,58 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-   <modelVersion>4.0.0</modelVersion>
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
 
-   <parent>
-      <groupId>org.picketlink</groupId>
-      <artifactId>picketlink-idm-parent</artifactId>
-      <version>3.0.0-SNAPSHOT</version>
-      <relativePath>../pom.xml</relativePath>
-   </parent>
+	<parent>
+		<groupId>org.picketlink</groupId>
+		<artifactId>picketlink-idm-parent</artifactId>
+		<version>3.0.0-SNAPSHOT</version>
+		<relativePath>../pom.xml</relativePath>
+	</parent>
 
-   <artifactId>picketlink-idm-api</artifactId>
-   <packaging>jar</packaging>
+	<artifactId>picketlink-idm-api</artifactId>
+	<packaging>jar</packaging>
 
-   <name>PicketLink Identity Management API</name>
+	<name>PicketLink Identity Management API</name>
 
-   <dependencies>
-      <dependency>
-         <groupId>org.hibernate.javax.persistence</groupId>
-         <artifactId>hibernate-jpa-2.0-api</artifactId>
-         <scope>provided</scope>
-      </dependency>
-   </dependencies>
+	<dependencies>
+		<dependency>
+			<groupId>org.hibernate.javax.persistence</groupId>
+			<artifactId>hibernate-jpa-2.0-api</artifactId>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.picketlink</groupId>
+			<artifactId>picketlink-common</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+	</dependencies>
 
-   <build>
-      <plugins>
+	<build>
+		<plugins>
 
-         <!-- Enable checkstyle execution -->
-         <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-checkstyle-plugin</artifactId>
-            <version>2.9.1</version>
-            <configuration>
-               <configLocation>../checkstyle.xml</configLocation>
-               <consoleOutput>true</consoleOutput>
-               <failsOnError>true</failsOnError>
-               <useFile />
-            </configuration>
-            <executions>
-               <execution>
-                  <id>check-style</id>
-                  <phase>compile</phase>
-                  <goals>
-                     <goal>checkstyle</goal>
-                  </goals>
-               </execution>
-            </executions>
-         </plugin>
-      </plugins>
-   </build>
+			<!-- Enable checkstyle execution -->
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-checkstyle-plugin</artifactId>
+				<version>2.9.1</version>
+				<configuration>
+					<configLocation>../checkstyle.xml</configLocation>
+					<consoleOutput>true</consoleOutput>
+					<failsOnError>true</failsOnError>
+					<useFile />
+				</configuration>
+				<executions>
+					<execution>
+						<id>check-style</id>
+						<phase>compile</phase>
+						<goals>
+							<goal>checkstyle</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
 
 </project>
