+++ idm/api/src/main/java/org/picketlink/idm/credential/Digest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.picketlink.idm.credential.internal;
+package org.picketlink.idm.credential;
 
 import org.picketlink.common.util.Base64;
