+++ idm/api/src/main/java/org/picketlink/idm/credential/X509Cert.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.picketlink.idm.credential.internal;
+package org.picketlink.idm.credential;
 
 import java.security.cert.X509Certificate;
