+++ idm/tests/src/test/java/org/picketlink/test/idm/credential/DigestCredentialTestCase.java
@@ -25,9 +25,9 @@ import java.util.Date;
 
 import org.junit.Test;
 import org.picketlink.idm.IdentityManager;
+import org.picketlink.idm.credential.Digest;
+import org.picketlink.idm.credential.DigestCredentials;
 import org.picketlink.idm.credential.Credentials.Status;
-import org.picketlink.idm.credential.internal.Digest;
-import org.picketlink.idm.credential.internal.DigestCredentials;
 import org.picketlink.idm.credential.internal.DigestUtil;
 import org.picketlink.idm.model.User;
 import org.picketlink.test.idm.AbstractIdentityManagerTestCase;
